Remove unnecessary reporter subclasses from public API
The `BenchmarkReporter` class is meant to be part of the public API so
custom reporters can be built. However, there isn't a good reason to
expose the `ConsoleReporter`, `CSVReporter`, or `JSONReporter` classes.
