+++ src/benchmark.cc
@@ -42,7 +42,9 @@
 #include "commandlineflags.h"
 #include "complexity.h"
 #include "counter.h"
+#include "csv_reporter.h"
 #include "internal_macros.h"
+#include "json_reporter.h"
 #include "log.h"
 #include "mutex.h"
 #include "perf_counters.h"
